+++ src/program/resolve/index.rs
@@ -722,6 +722,7 @@ mod tests {
             param_sig_key: String::new(),
             return_type: None,
             return_type_id: None,
+            abi_overload_collapsed: false,
         }
     }
 
@@ -751,6 +752,7 @@ mod tests {
             param_sig_key: String::new(),
             return_type: None,
             return_type_id: None,
+            abi_overload_collapsed: false,
         }
     }
 
@@ -781,6 +783,7 @@ mod tests {
             param_sig_key: String::new(),
             return_type: None,
             return_type_id: None,
+            abi_overload_collapsed: false,
         }
     }